Replace priority queue with limited queue to prevent OOM slow polling issue
