+++ sonic_data_client/mixed_db_client.go
@@ -99,16 +99,6 @@ var RedisDbMap map[string]*redis.Client = nil
 // Db num from database configuration
 var DbInstNum = 0
 
-func enqueFatalMsgMixed(c *MixedDbClient, msg string) {
-	log.Error(msg)
-	c.q.ForceEnqueueItem(Value{
-		&spb.Value{
-			Timestamp: time.Now().UnixNano(),
-			Fatal:     msg,
-		},
-	})
-}
-
 func Hget(configDbConnector *swsscommon.ConfigDBConnector, table string, key string, field string) (string, error) {
 	var fieldValuePairs = configDbConnector.Get_entry(table, key)
 	defer swsscommon.DeleteFieldValueMap(fieldValuePairs)
@@ -1701,7 +1691,7 @@ func (c *MixedDbClient) StreamRun(q *LimitedQueue, stop chan struct{}, w *sync.W
 				c.synced.Add(1)
 				go c.streamOnChangeSubscription(sub.GetPath())
 			} else {
-				enqueFatalMsgMixed(c, fmt.Sprintf("unsupported subscription mode, %v", subMode))
+				enqueFatalMsg(c.q, fmt.Sprintf("unsupported subscription mode, %v", subMode))
 				return
 			}
 		}
@@ -1728,7 +1718,7 @@ func (c *MixedDbClient) streamOnChangeSubscription(gnmiPath *gnmipb.Path) {
 	tblPaths, err := c.getDbtablePath(gnmiPath, nil)
 	if err != nil {
 		msg := fmt.Sprintf("streamOnChangeSubscription error:  %v", err)
-		enqueFatalMsgMixed(c, msg)
+		enqueFatalMsg(c.q, msg)
 		c.synced.Done()
 		c.w.Done()
 		return
@@ -1747,7 +1737,7 @@ func (c *MixedDbClient) streamOnChangeSubscription(gnmiPath *gnmipb.Path) {
 func (c *MixedDbClient) streamSampleSubscription(sub *gnmipb.Subscription, updateOnly bool) {
 	samplingInterval, err := validateSampleInterval(sub)
 	if err != nil {
-		enqueFatalMsgMixed(c, err.Error())
+		enqueFatalMsg(c.q, err.Error())
 		c.synced.Done()
 		c.w.Done()
 		return
@@ -1756,7 +1746,7 @@ func (c *MixedDbClient) streamSampleSubscription(sub *gnmipb.Subscription, updat
 	gnmiPath := sub.GetPath()
 	tblPaths, err := c.getDbtablePath(gnmiPath, nil)
 	if err != nil {
-		enqueFatalMsgMixed(c, err.Error())
+		enqueFatalMsg(c.q, err.Error())
 		c.synced.Done()
 		c.w.Done()
 		return
@@ -1778,7 +1768,7 @@ func (c *MixedDbClient) dbFieldSubscribe(gnmiPath *gnmipb.Path, onChange bool, i
 
 	tblPaths, err := c.getDbtablePath(gnmiPath, nil)
 	if err != nil {
-		enqueFatalMsgMixed(c, err.Error())
+		enqueFatalMsg(c.q, err.Error())
 		c.synced.Done()
 		return
 	}
@@ -1787,7 +1777,7 @@ func (c *MixedDbClient) dbFieldSubscribe(gnmiPath *gnmipb.Path, onChange bool, i
 	redisDb, ok := RedisDbMap[c.mapkey+":"+tblPath.dbName]
 	if !ok {
 		msg := fmt.Sprintf("RedisDbMap not exist:  %v", c.mapkey+":"+tblPath.dbName)
-		enqueFatalMsgMixed(c, msg)
+		enqueFatalMsg(c.q, msg)
 		c.synced.Done()
 		return
 	}
@@ -1836,7 +1826,7 @@ func (c *MixedDbClient) dbFieldSubscribe(gnmiPath *gnmipb.Path, onChange bool, i
 	val := readVal()
 	err = sendVal(val)
 	if err != nil {
-		enqueFatalMsgMixed(c, err.Error())
+		enqueFatalMsg(c.q, err.Error())
 		c.synced.Done()
 		return
 	}
@@ -1913,7 +1903,7 @@ func (c *MixedDbClient) dbSingleTableKeySubscribe(rsd redisSubData, updateChanne
 				if tblPath.tableKey != "" {
 					err = c.tableData2Msi(&tblPath, false, nil, &newMsi)
 					if err != nil {
-						enqueFatalMsgMixed(c, err.Error())
+						enqueFatalMsg(c.q, err.Error())
 						return
 					}
 				} else {
@@ -1925,7 +1915,7 @@ func (c *MixedDbClient) dbSingleTableKeySubscribe(rsd redisSubData, updateChanne
 					tblPath.tableKey = subscr.Channel[prefixLen:]
 					err = c.tableData2Msi(&tblPath, true, nil, &newMsi)
 					if err != nil {
-						enqueFatalMsgMixed(c, err.Error())
+						enqueFatalMsg(c.q, err.Error())
 						return
 					}
 				}
@@ -1971,7 +1961,7 @@ func (c *MixedDbClient) dbTableKeySubscribe(gnmiPath *gnmipb.Path, interval time
 	// Helper to handle fatal case.
 	handleFatalMsg := func(msg string) {
 		log.V(1).Infof(msg)
-		enqueFatalMsgMixed(c, msg)
+		enqueFatalMsg(c.q, msg)
 		signalSync()
 	}
